fix(core): Ensure ComponentFixture.detectChanges causes synchronous ApplicationRef.tick
When ZoneJS is enabled, `ComponentFixture.detectChanges` _may_ cause a
synchronous `ApplicationRef.tick` if the zone is stable after the
`ngZone.run` call. If the zone is not stable, `ApplicationRef.tick` will
not happen until it `onMicrotaskEmpty` emits. Render hooks execute
within `ApplicationRef.tick` so it is important that these are
predictably run when calling `ComponentFixture.detectChanges`.

This change completes the mering of behaviors between zone-based and
zoneless fixtures. The only difference now is the default value of
autoDetect, which is true for zoneless and false for zones.