Update animation_player.ts as requested
To address the confusion between the description of position and how it's being used, you should revise the documentation so it accurately reflects that position can be represented either as a 0-based millisecond value or a fractional value between 0 and 1. You also need to clarify when each representation is used.