fix(core): correct type narrowing on Signal<T>
TypeScript considers an intersection of two function types to be an
overloaded function containing both signatures. The implication of this
is that `Signal<T|undefined> & Signal<T>` is just
`Signal<T|undefined>`, instead of `Signal<T>` as would be expected.

This PR introduces a new symbol that holds the type information for
`Signal`, so that intersection types will calculate the type
intersection correctly, allowing for `WritableResource#hasValue` to
correctly narrow the type of `WritableResource#value`.